pyproject.toml
[tool.poetry]
name = "github-tag-summarizer"
version = "0.1.0"
description = ""
authors = ["Masahiro Nishimi <nishimi@sonicgarden.jp>"]
readme = "README.md"
packages = [{include = "github_tag_summarizer"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.0"
tiktoken = "^0.3.2"
langchain = "^0.0.123"
openai = "^0.27.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
